+++ files/k3s-install-server.sh
@@ -1,5 +1,28 @@
 #!/bin/bash
 
+check_os() {
+  name=$(cat /etc/os-release | grep ^NAME= | sed 's/"//g')
+  clean_name=$${name#*=}
+
+  version=$(cat /etc/os-release | grep ^VERSION_ID= | sed 's/"//g')
+  clean_version=$${version#*=}
+  major=$${clean_version%.*}
+  minor=$${clean_version#*.}
+  
+  if [[ "$clean_name" == "Ubuntu" ]]; then
+    operating_system="ubuntu"
+  elif [[ "$clean_name" == "Oracle Linux Server" ]]; then
+    operating_system="oraclelinux"
+  else
+    operating_system="undef"
+  fi
+
+  echo "K3S install process running on: "
+  echo "OS: $operating_system"
+  echo "OS Major Release: $major"
+  echo "OS Minor Release: $minor"
+}
+
 wait_lb() {
 while [ true ]
 do
@@ -13,7 +36,7 @@ done
 }
 
 render_nginx_config(){
-cat << 'EOF' > $NGINX_RESOURCES_FILE
+cat << 'EOF' > "$NGINX_RESOURCES_FILE"
 ---
 apiVersion: v1
 kind: Service
@@ -62,7 +85,7 @@ EOF
 
 render_staging_issuer(){
 STAGING_ISSUER_RESOURCE=$1
-cat << 'EOF' > $STAGING_ISSUER_RESOURCE
+cat << 'EOF' > "$STAGING_ISSUER_RESOURCE"
 apiVersion: cert-manager.io/v1
 kind: ClusterIssuer
 metadata:
@@ -87,7 +110,7 @@ EOF
 
 render_prod_issuer(){
 PROD_ISSUER_RESOURCE=$1
-cat << 'EOF' > $PROD_ISSUER_RESOURCE
+cat << 'EOF' > "$PROD_ISSUER_RESOURCE"
 apiVersion: cert-manager.io/v1
 kind: ClusterIssuer
 metadata:
@@ -110,47 +133,97 @@ spec:
 EOF
 }
 
-# Disable firewall 
-/usr/sbin/netfilter-persistent stop
-/usr/sbin/netfilter-persistent flush
+check_os
 
-systemctl stop netfilter-persistent.service
-systemctl disable netfilter-persistent.service
-# END Disable firewall
+if [[ "$operating_system" == "ubuntu" ]]; then
+  echo "Canonical Ubuntu"
+  # Disable firewall 
+  /usr/sbin/netfilter-persistent stop
+  /usr/sbin/netfilter-persistent flush
 
-apt-get update
-apt-get install -y software-properties-common jq
-DEBIAN_FRONTEND=noninteractive apt-get upgrade -y
-DEBIAN_FRONTEND=noninteractive apt-get install --no-install-recommends -y  python3 python3-pip
-pip install oci-cli
+  systemctl stop netfilter-persistent.service
+  systemctl disable netfilter-persistent.service
+  # END Disable firewall
 
-local_ip=$(curl -s -H "Authorization: Bearer Oracle" -L http://169.254.169.254/opc/v2/vnics/ | jq -r '.[0].privateIp')
-flannel_iface=$(ip -4 route ls | grep default | grep -Po '(?<=dev )(\S+)')
+  apt-get update
+  apt-get install -y software-properties-common jq
+  DEBIAN_FRONTEND=noninteractive apt-get upgrade -y
+  DEBIAN_FRONTEND=noninteractive apt-get install --no-install-recommends -y  python3 python3-pip
+  pip install oci-cli
+
+  # Fix /var/log/journal dir size
+  echo "SystemMaxUse=100M" >> /etc/systemd/journald.conf
+  echo "SystemMaxFileSize=100M" >> /etc/systemd/journald.conf
+  systemctl restart systemd-journald
+fi
+
+if [[ "$operating_system" == "oraclelinux" ]]; then
+  echo "Oracle Linux"
+  # Disable firewall
+  systemctl disable --now firewalld
+  # END Disable firewall
+
+  # Fix iptables/SELinux bug
+  echo '(allow iptables_t cgroup_t (dir (ioctl)))' > /root/local_iptables.cil
+  semodule -i /root/local_iptables.cil
+
+  dnf -y update
+
+  if [[ $major -eq 9 ]]; then
+    dnf -y install oraclelinux-developer-release-el9
+    dnf -y install jq python39-oci-cli curl
+  else
+    dnf -y install oraclelinux-developer-release-el8
+    dnf -y module enable python36:3.6
+    dnf -y install jq python36-oci-cli curl
+  fi
+fi
 
 export OCI_CLI_AUTH=instance_principal
 first_instance=$(oci compute instance list --compartment-id ${compartment_ocid} --availability-domain ${availability_domain} --lifecycle-state RUNNING --sort-by TIMECREATED  | jq -r '.data[]|select(."display-name" | endswith("k3s-servers")) | .["display-name"]' | tail -n 1)
 instance_id=$(curl -s -H "Authorization: Bearer Oracle" -L http://169.254.169.254/opc/v2/instance | jq -r '.displayName')
-first_last="last"
+
+k3s_install_params=("--tls-san ${k3s_tls_san}")
+
+%{ if k3s_subnet != "default_route_table" } 
+local_ip=$(ip -4 route ls ${k3s_subnet} | grep -Po '(?<=src )(\S+)')
+flannel_iface=$(ip -4 route ls ${k3s_subnet} | grep -Po '(?<=dev )(\S+)')
+
+k3s_install_params+=("--node-ip $local_ip")
+k3s_install_params+=("--advertise-address $local_ip")
+k3s_install_params+=("--flannel-iface $flannel_iface")
+%{ endif }
 
 %{ if install_nginx_ingress } 
-disable_traefik="--disable traefik"
+k3s_install_params+=("--disable traefik")
 %{ endif }
 
 %{ if expose_kubeapi }
-tls_extra_san="--tls-san ${k3s_tls_san_public}"
+k3s_install_params+=("--tls-san ${k3s_tls_san_public}")
+%{ endif }
+
+if [[ "$operating_system" == "oraclelinux" ]]; then
+  k3s_install_params+=("--selinux")
+fi
+
+INSTALL_PARAMS="$${k3s_install_params[*]}"
+
+%{ if k3s_version == "latest" }
+K3S_VERSION=$(curl --silent https://api.github.com/repos/k3s-io/k3s/releases/latest | jq -r '.name')
+%{ else }
+K3S_VERSION="${k3s_version}"
 %{ endif }
 
 if [[ "$first_instance" == "$instance_id" ]]; then
   echo "I'm the first yeeee: Cluster init!"
-  first_last="first"
-  until (curl -sfL https://get.k3s.io | K3S_TOKEN=${k3s_token} sh -s - --cluster-init $disable_traefik --node-ip $local_ip --advertise-address $local_ip --flannel-iface $flannel_iface --tls-san ${k3s_tls_san} $tls_extra_san); do
+  until (curl -sfL https://get.k3s.io | INSTALL_K3S_VERSION=$K3S_VERSION K3S_TOKEN=${k3s_token} sh -s - --cluster-init $INSTALL_PARAMS); do
     echo 'k3s did not install correctly'
     sleep 2
   done
 else
   echo ":( Cluster join"
   wait_lb
-  until (curl -sfL https://get.k3s.io | K3S_TOKEN=${k3s_token} sh -s - --server https://${k3s_url}:6443 $disable_traefik --node-ip $local_ip --advertise-address $local_ip --flannel-iface $flannel_iface --tls-san ${k3s_tls_san} $tls_extra_san); do
+  until (curl -sfL https://get.k3s.io | INSTALL_K3S_VERSION=$K3S_VERSION K3S_TOKEN=${k3s_token} sh -s - --server https://${k3s_url}:6443 $INSTALL_PARAMS); do
     echo 'k3s did not install correctly'
     sleep 2
   done
@@ -163,17 +236,18 @@ until kubectl get pods -A | grep 'Running'; do
 done
 
 %{ if install_longhorn }
-if [[ "$first_last" == "first" ]]; then
-  DEBIAN_FRONTEND=noninteractive apt-get install --no-install-recommends -y  open-iscsi curl util-linux
-  systemctl enable iscsid.service
-  systemctl start iscsid.service
+if [[ "$first_instance" == "$instance_id" ]]; then
+  if [[ "$operating_system" == "ubuntu" ]]; then
+    DEBIAN_FRONTEND=noninteractive apt-get install --no-install-recommends -y  open-iscsi curl util-linux
+  fi
+
+  systemctl enable --now iscsid.service
   kubectl apply -f https://raw.githubusercontent.com/longhorn/longhorn/${longhorn_release}/deploy/longhorn.yaml
-  kubectl create -f https://raw.githubusercontent.com/longhorn/longhorn/${longhorn_release}/examples/storageclass.yaml
 fi
 %{ endif }
 
 %{ if install_nginx_ingress }
-if [[ "$first_last" == "first" ]]; then
+if [[ "$first_instance" == "$instance_id" ]]; then
   kubectl apply -f https://raw.githubusercontent.com/kubernetes/ingress-nginx/controller-${nginx_ingress_release}/deploy/static/provider/baremetal/deploy.yaml
   NGINX_RESOURCES_FILE=/root/nginx-ingress-resources.yaml
   render_nginx_config
@@ -182,7 +256,7 @@ fi
 %{ endif }
 
 %{ if install_certmanager }
-if [[ "$first_last" == "first" ]]; then
+if [[ "$first_instance" == "$instance_id" ]]; then
   kubectl apply -f https://github.com/cert-manager/cert-manager/releases/download/${certmanager_release}/cert-manager.yaml
   render_staging_issuer /root/staging_issuer.yaml
   render_prod_issuer /root/prod_issuer.yaml
